// src/schemas/vnu-jar.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type VNuMessageObject =
    | {
          type: 'info';
          subType?: 'warning';
          message?: string;
          extract?: string;
          offset?: number;
          url?: string;
          firstLine?: number;
          firstColumn?: number;
          lastLine?: number;
          lastColumn?: number;
          hiliteStart?: number;
          hiliteLength?: number;
      }
    | {
          type: 'error';
          subType?: 'fatal';
          message?: string;
          extract?: string;
          offset?: number;
          url?: string;
          firstLine?: number;
          firstColumn?: number;
          lastLine?: number;
          lastColumn?: number;
          hiliteStart?: number;
          hiliteLength?: number;
      }
    | {
          type: 'non-document-error';
          subType?: 'io' | 'schema' | 'internal';
          message?: string;
          extract?: string;
          offset?: number;
          url?: string;
          firstLine?: number;
          firstColumn?: number;
          lastLine?: number;
          lastColumn?: number;
          hiliteStart?: number;
          hiliteLength?: number;
      };

export interface VNuJSONSchema {
    messages: VNuMessageObject[];
    url?: string;
    source?: VNuSourceObject;
    language?: string;
}
export interface VNuSourceObject {
    code: string;
    type?: string;
    encoding?: string;
}